Fixed to get list of ets for Owners.

+++ Petshop.API.UI/Controllers/OwnerController.cs
@@ -31,7 +31,7 @@ namespace Petshop.RestAPI.UI.Controllers
 
         // GET api/<OwnerController>/5
         [HttpGet("{id}")]
-        public Owner Get(int id)
+        public ActionResult<Owner> Get(int id)
         {
             return _ownerService.FindOwnerByID(id);
         }
@@ -47,31 +47,50 @@ namespace Petshop.RestAPI.UI.Controllers
         }
         // POST api/<OwnerController>
         [HttpPost]
-        public Owner Post([FromBody] NewOwnerObject theOwner)
+        public ActionResult<Owner> Post([FromBody] NewOwnerObject theOwner)
         {
+            if(string.IsNullOrEmpty(theOwner.FirstName) || string.IsNullOrEmpty(theOwner.LastName) || string.IsNullOrEmpty(theOwner.Address) || string.IsNullOrEmpty(theOwner.PhoneNr) || string.IsNullOrEmpty(theOwner.Email))
+            {
+                return BadRequest("You have not entered all the needed data.");
+            }
             return _ownerService.AddNewOwner(theOwner.FirstName, theOwner.LastName, theOwner.Address, theOwner.PhoneNr, theOwner.Email);
         }
 
         public class updateOwnerObj
         {
-            public int updateParam { get; set; }
+            public int? updateParam { get; set; }
             public string updateData { get; set; }
         }
 
         // PUT api/<OwnerController>/5
         [HttpPut("{id}")]
-        public Owner Put(int id, [FromBody] updateOwnerObj theOwnerObj)
+        public ActionResult<Owner> Put(int id, [FromBody] updateOwnerObj theOwnerObj)
         {
             Owner theOwner = _ownerService.FindOwnerByID(id);
-            return _ownerService.UpdateOwner(theOwner, theOwnerObj.updateParam, theOwnerObj.updateData);
+            if (theOwner == null)
+            {
+                return NotFound("Could not find that owner to update.");
+            }
+
+            if (theOwnerObj.updateParam == null || string.IsNullOrEmpty(theOwnerObj.updateData))
+            {
+                return BadRequest("You have not entered the necessary data.");
+            }
+            
+            return _ownerService.UpdateOwner(theOwner, theOwnerObj.updateParam.Value, theOwnerObj.updateData);
         }
 
         // DELETE api/<OwnerController>/5
         [HttpDelete("{id}")]
-        public string Delete(int id)
+        public ActionResult Delete(int id)
         {
+            Owner toDelete = _ownerService.FindOwnerByID(id);
+            if(toDelete == null)
+            {
+                return NotFound("Sorry could not find that owner to delete.");
+            }
             _ownerService.DeleteOwnerByID(id);
-            return "Owner deleted.";
+            return Ok($"Owner with Id {id} deleted.");
         }
     }
 }
